+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- General
+  - Use lower-case for refer/import/require sorting. #560 #561
+
 - Editor
   - Fix `resolve macro as` code action after regression introduced recently.
   - Fix `unused-public-var` not being suppressed during project startup. #554